feat: change the return data type of fibonacci calculate to long

+++ src/main/java/cc/xpbootcamp/warmup/fibonacci/Fibonacci.java
@@ -2,7 +2,7 @@ package cc.xpbootcamp.warmup.fibonacci;
 
 public class Fibonacci {
 
-    public int calculate(int position) {
+    public long calculate(int position) {
         if (position == 1 || position == 2) {
             return 1;
         }
